Extract `FeedLoaderStub` into its own file
